corrected the int16/int32 problem in hypervectors

+++ Hypervectors/include/random_projection.h
@@ -5,6 +5,6 @@
 #include <vector>
 #include <cstdint>
 
-std::vector<int32_t> transform_set_into_vector(const std::unordered_set<unsigned long int> &hashes, int d);
+std::vector<int16_t> transform_set_into_vector(const std::unordered_set<unsigned long int> &hashes, int d);
 
 #endif // RANDOM_PROJECTION_H
